+++ lib/enums/timelineEvents.d.ts
@@ -10,6 +10,7 @@ export declare enum TimelineEvents {
     DragFinished = "dragFinished",
     Scroll = "scroll",
     ScrollFinished = "scrollFinished",
+    ContextMenu = "onContextMenu",
     DoubleClick = "doubleClick",
     MouseDown = "mouseDown"
 }