+++ lib/settings/styles/timelineCapStyle.d.ts
@@ -14,11 +14,11 @@ export interface TimelineCapStyle {
     /**
      * Cap stroke color.
      */
-    strokeColor: string;
+    strokeColor?: string;
     /**
      * Cap fill color.
      */
-    fillColor: string;
+    fillColor?: string;
     /**
      * Cap type
      */